+++ CHANGELOG.rst
@@ -24,6 +24,8 @@ Breaking Changes
 ^^^^^^^^^^^^^^^^
 
 * Moved ``client.flow_documents`` to ``client.flow.documents``
+* ``BaseEntity.set_attributes()`` from ``airslate.entities.base`` module has been
+  removed.
 
 
 Trivial/Internal Changes